🔧 Migrate deprecated pytest fixtures (#7623)
Remove the `aiida.manage.tests.pytest_fixtures` module, which has been
deprecated since aiida-core==2.6 and was scheduled for removal in 3.0,
together with its dedicated test `.github/test_legacy_pytest_fixtures.py`
and the corresponding CI step. The module's `EntryPointManager` lives on
in `aiida.tools.pytest_fixtures.entry_points`.

Plugin packages should load the replacement fixtures with:

    pytest_plugins = 'aiida.tools.pytest_fixtures'

Daemon tests have been migrated to new pytest fixture tests. Docs and
the changelog are updated accordingly.
